refactor: оптимизировать классы параметров поиска Events и Comments
