+++ main-service/src/main/java/ru/practicum/mainservice/comment/service/CommentService.java
@@ -12,8 +12,8 @@ import ru.practicum.mainservice.comment.mapper.CommentMapper;
 import ru.practicum.mainservice.comment.model.Comment;
 import ru.practicum.mainservice.comment.model.QComment;
 import ru.practicum.mainservice.comment.repository.CommentRepository;
-import ru.practicum.mainservice.controllers.admincontrollers.parameters.AdminCommentRequestParameters;
-import ru.practicum.mainservice.controllers.parameters.BaseCommentRequestParameters;
+import ru.practicum.mainservice.controllers.parameters.comment.AdminCommentRequestParameters;
+import ru.practicum.mainservice.controllers.parameters.comment.BaseCommentRequestParameters;
 import ru.practicum.mainservice.event.model.Event;
 import ru.practicum.mainservice.event.model.QEvent;
 import ru.practicum.mainservice.event.repository.EventRepository;
