+++ main-service/src/main/java/ru/practicum/mainservice/controllers/admincontrollers/CommentAdminController.java
@@ -8,7 +8,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.practicum.mainservice.comment.dto.CommentDto;
 import ru.practicum.mainservice.comment.service.CommentService;
-import ru.practicum.mainservice.controllers.admincontrollers.parameters.AdminCommentRequestParameters;
+import ru.practicum.mainservice.controllers.parameters.comment.AdminCommentRequestParameters;
 
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.PositiveOrZero;
