+++ main-service/src/main/java/ru/practicum/mainservice/controllers/admincontrollers/EventAdminController.java
@@ -6,7 +6,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.practicum.mainservice.controllers.admincontrollers.parameters.EventAdminRequestParameters;
+import ru.practicum.mainservice.controllers.parameters.event.EventAdminRequestParameters;
 import ru.practicum.mainservice.event.dto.EventFullDto;
 import ru.practicum.mainservice.event.dto.UpdateEventAdminRequest;
 import ru.practicum.mainservice.event.model.EventState;
