+++ main-service/src/main/java/ru/practicum/mainservice/controllers/parameters/comment/AdminCommentRequestParameters.java
@@ -1,7 +1,8 @@
-package ru.practicum.mainservice.controllers.admincontrollers.parameters;
+package ru.practicum.mainservice.controllers.parameters.comment;
 
 import lombok.*;
-import ru.practicum.mainservice.controllers.parameters.BaseCommentRequestParameters;
+import lombok.experimental.SuperBuilder;
+import ru.practicum.mainservice.controllers.parameters.comment.BaseCommentRequestParameters;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -11,6 +12,7 @@ import java.util.List;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
+@SuperBuilder
 public class AdminCommentRequestParameters extends BaseCommentRequestParameters {
     private List<Long> userIds;
