+++ main-service/src/main/java/ru/practicum/mainservice/controllers/parameters/event/EventAdminRequestParameters.java
@@ -1,9 +1,8 @@
-package ru.practicum.mainservice.controllers.admincontrollers.parameters;
+package ru.practicum.mainservice.controllers.parameters.event;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
-import ru.practicum.mainservice.controllers.parameters.BaseEventRequestParameters;
 import ru.practicum.mainservice.event.model.EventState;
 
 import java.util.List;
