+++ main-service/src/main/java/ru/practicum/mainservice/controllers/publiccontrollers/EventPublicController.java
@@ -9,8 +9,8 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import ru.practicum.mainservice.controllers.publiccontrollers.parameters.EventPublicRequestParameters;
-import ru.practicum.mainservice.controllers.publiccontrollers.parameters.EventRequestSort;
+import ru.practicum.mainservice.controllers.parameters.event.EventPublicRequestParameters;
+import ru.practicum.mainservice.controllers.parameters.event.EventRequestSort;
 import ru.practicum.mainservice.event.dto.EventFullDto;
 import ru.practicum.mainservice.event.dto.EventShortDto;
 import ru.practicum.mainservice.event.service.EventService;
